+++ src/base/lib/Convex/PlutusLedger.hs
@@ -53,6 +53,8 @@ module Convex.PlutusLedger(
 
   -- * Value
   unTransTxOutValue,
+
+  transValue,
   unTransValue,
 
   -- * Scripts
@@ -86,6 +88,7 @@ import           Plutus.V1.Ledger.Interval (Closure, Extended (..),
                                             UpperBound (..))
 import qualified Plutus.V1.Ledger.Scripts  as P
 import qualified Plutus.V1.Ledger.Value    as Value
+import qualified PlutusTx.AssocMap         as Map
 import qualified PlutusTx.Prelude          as PlutusTx
 
 transScriptHash :: C.ScriptHash -> PV1.ValidatorHash
@@ -217,6 +220,13 @@ unTransValue =
   where
     toSingleton (cs, tn, q) =
         unTransAssetId (Value.assetClass cs tn) <&> (, C.Quantity q)
+
+transValue :: C.Value -> PV1.Value
+transValue =
+  let t (assetId, C.Quantity quantity) =
+        let Value.AssetClass (sym, tn) = transAssetId assetId
+        in (sym, Map.singleton tn quantity)
+  in PV1.Value . Map.fromList . fmap t . C.valueToList
 
 unTransPlutusScript
     :: C.SerialiseAsRawBytes plutusScript