+++ changelog.md
@@ -27,7 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * `convex-base`
-  - `Convex.PlutusLedger`: `transAddressInEra`, prisms and isos for the `Interval` type
+  - `Convex.PlutusLedger`: `transAddressInEra`, prisms and isos for the `Interval` type, `transValue`
   - `Convex.Utxos.singleton`
   - `Convex.Utxos`: Added redeemer and ex units to `RemoveUtxoEvent`
   - `Convex.Lenses`: Added `_TxOutDatumInline`, `_TxOutDatumInTx`, `_ScriptData`, `_PlutusScriptWitness`, `_TxExtraKeyWitnesses`, prisms for validity intervals, `_TxOutDatumHash`